+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [0.1.1](https://github.com/Blobfolio/trimothy/releases/tag/v0.1.1) - 2022-04-11
 
+## Changes
+
+* Minor performance improvements
+
 ## Fixed
 
-* Return empty slice on non-empty, all match.
+* Return empty slice when all bytes match trim predicate.
 * Markdown (docs) formatting issues.